MappedMemory has been completely removed from all mappers

+++ crates/nes/src/machine.rs
@@ -10,7 +10,7 @@ use crate::cpu::{Cpu, CpuPinIn, TickResult};
 use crate::debug::{Debug, DebugEvent};
 use crate::input::Input;
 use crate::mapper::{RcMapper, SaveWram};
-use crate::memory::MemoryBlock;
+use crate::memory::{Memory, MemoryBlock};
 use crate::ppu::{FrameEnd, Ppu};
 use crate::region::Region;
 
@@ -241,7 +241,7 @@ impl Machine {
 
     fn read(&mut self, addr: u16) -> u8 {
         let value = match self.cpu_bus.read_addr(addr) {
-            Some((addr, DeviceKind::CpuRam)) => self.cpu_mem.read(addr),
+            Some((addr, DeviceKind::CpuRam)) => self.cpu_mem.read(addr as usize),
             Some((addr, DeviceKind::Ppu)) => self.ppu.read(addr),
             Some((addr, DeviceKind::Mapper)) => self.mapper.read(BusKind::Cpu, addr),
             Some((addr, DeviceKind::Input)) => self.input.read(addr, self.cpu_bus.open_bus.get()),
@@ -263,7 +263,7 @@ impl Machine {
         // Loop through potential mappings to allow MMC5 to snoop on PPU register writes
         for mapping in self.cpu_bus.write_addrs(addr) {
             match mapping {
-                (addr, DeviceKind::CpuRam) => self.cpu_mem.write(addr, value),
+                (addr, DeviceKind::CpuRam) => self.cpu_mem.write(addr as usize, value),
                 (addr, DeviceKind::Ppu) => self.ppu.write(addr, value),
                 (addr, DeviceKind::Mapper) => self.mapper.write(BusKind::Cpu, addr, value),
                 (addr, DeviceKind::Input) => self.input.write(addr, value),
@@ -281,7 +281,7 @@ impl Machine {
     #[cfg(feature = "debugger")]
     pub fn peek(&self, addr: u16) -> u8 {
         match self.cpu_bus.read_addr(addr) {
-            Some((addr, DeviceKind::CpuRam)) => self.cpu_mem.read(addr),
+            Some((addr, DeviceKind::CpuRam)) => self.cpu_mem.read(addr as usize),
             Some((addr, DeviceKind::Ppu)) => self.ppu.peek(addr),
             Some((addr, DeviceKind::Mapper)) => self.mapper.peek(BusKind::Cpu, addr),
             Some((addr, DeviceKind::Input)) => self.input.peek(addr, self.cpu_bus.open_bus.get()),
